Merged successful branch with main (I think) but some big progress in this commit:
Used the constructor of gridGenerator to dynamically request a certain size of grid dependent on the ship sizes in the Ships enum class, and then dynamically prevented the user from entering a grid that would lead to an infinite for loop (see text discussion in local files)

Randomly determined whether the ships should be placed horizontally or vertically.

Next commit:

Allow the user to place ships and then fight against an AI, after this we call it.